rename package + classes restrct
