+++ PenBot/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>robo.testbot.pen</name>
+	<name>PenBot</name>
 	<comment></comment>
 	<projects>
 	</projects>
