+++ PenBot/build.properties
@@ -1,8 +1,8 @@
 nxj.home=${env.NXJ_HOME}
 #nxj.home=../snapshot
 
-main.class=PenBot
-package=robo.testbot.pen
+main.class=Main
+package=penbot
 
 docs=doc
